requirements.txt
flask
pymupdf
ocrmypdf
python-docx
